Make zk-atoms actually persist
I spoke too soon before when I said this wasn't an issue. It turns out
that zk-atoms would persist as long as _some_ client was
connected, but otherwise they would disappear. Weirdly, this isn't a
problem with zk-refs.

This commit changes the way that zk-atoms are created in ZooKeeper by
adding `:persistent? true`.